Move locale-switching code in test to a shared header
[On dev, this is also used by QString, which originally motivated the
move; but improvements made in the move are worth including in 6.2,
and moving to the shared header is the least disruptive way to adapt
the cherry-pick, even though tst_qstring.cpp needs no change now.]

In the process, I noticed that setlocale() only returns the prior
value when passed nullptr as the new value; so rework the
implementation to get that right, so that it now correctly restores
the prior locale. That, in turn, means there's now a later call to
setlocale(), when we actually set the changed setting, which may
invalidate the earlier return; so copy it to a QByteArray before the
second call.

Included Ivan Solovev's improved version of how to reset the locale,
since TransientLocale needs it.

Change-Id: I4cb1efbda42f0e2cdd934e04b3b3732ce0f45a06
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 458d53f5723314b946d0b401bf91e417131263e9)